requirements.txt
cycler==0.10.0
linecache2==1.0.0
marshmallow==2.13.5
matplotlib==2.0.2
nose==1.3.7
numpy==1.13.1
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
rectpack==0.2
six==1.10.0
traceback2==1.4.0
unittest2==1.1.0
uvloop==0.8.0
